Create Networking Fundamentals
Networking Fundamentals

Played the role of a network security analyst.  I was tasked with completing a network vulnerability assessment of the Hollywood RockStar Corp office location.  I also had to complete several steps in order to analyze the Hollywood network and then provide RockStar Corp a summary of your findings.  The last part of the task was to determine if a suspected malicious hacker at the company anything behind that is suspicious and present my findings.